Styling is starting to look like something...
I have spent the better part of a day playing around with the best approach to making a vertical slider. And my conclusion is, there is none.. So I have gone with rotating the element and calculating the placement, so it will be the height of the card and all the way to the right.
The text has also been styled a tiny amount. You can change between forcing the text to be one line or it can line-break if it's too long.

Now just for the functionality.

Side note:
I have decided not to make this as all-around hybrid card, the code would be too messy.
So I will instead be making a series of my-button cards. This one will be "my-button-light", the next one will probably be "my-button-cover" or "my-button-tv".

This is all the customization that's possible now:
```yaml
  - type: custom:my-button
    entity: light.grono_lamp_office
    styles:
      card:
        height: 100%
        width: 100%
        borderRadius: 10%
        backgroundColor: rgba(255, 255, 255, 1)
        backgroundColorOff: rgba(230, 230, 230, 0.5)
      slider:
        position: "right"
        height: 40px
        backgroundColor: transparent
        progressColor: green
      text:
        row1:
          color: black
          oneLine: true
          size: 15px
        row2:
          color: black
          oneLine: true
          size: 13px
        row3:
          color: black
          oneLine: true
          size: 13px
    tap_action:
      action: toggle
```